+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDRenderPipelineRuntimeResources.cs
@@ -309,8 +309,10 @@ namespace UnityEngine.Rendering.HighDefinition
             public Shader temporalAntialiasingPS;
             [Reload("Runtime/PostProcessing/Shaders/UpsampleScene.compute")]
             public ComputeShader upsampleSceneCS;
+            [Reload("Runtime/PostProcessing/Shaders/LensFlareDataDriven.shader")]
+            public Shader lensFlareDataDrivenPS;
 
-            // Physicaly based DoF
+            // Physically based DoF
             [Reload("Runtime/PostProcessing/Shaders/DoFCircleOfConfusion.compute")]
             public ComputeShader dofCircleOfConfusion;
             [Reload("Runtime/PostProcessing/Shaders/DoFGather.compute")]