+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDStringConstants.cs
@@ -794,6 +794,15 @@ namespace UnityEngine.Rendering.HighDefinition
         public static readonly int _InputNearAlphaTexture          = Shader.PropertyToID("_InputNearAlphaTexture");
         public static readonly int _CoCTargetScale                 = Shader.PropertyToID("_CoCTargetScale");
 
+        public static readonly int _FlareTex                       = Shader.PropertyToID("_FlareTex");
+        public static readonly int _FlareColorValue                = Shader.PropertyToID("_FlareColorValue");
+        public static readonly int _FlareData0                     = Shader.PropertyToID("_FlareData0");
+        public static readonly int _FlareData1                     = Shader.PropertyToID("_FlareData1");
+        public static readonly int _FlareData2                     = Shader.PropertyToID("_FlareData2");
+        public static readonly int _FlareData3                     = Shader.PropertyToID("_FlareData3");
+        public static readonly int _FlareData4                     = Shader.PropertyToID("_FlareData4");
+        public static readonly int _FlareData5                     = Shader.PropertyToID("_FlareData5");
+
         public static readonly int _BloomParams                    = Shader.PropertyToID("_BloomParams");
         public static readonly int _BloomTint                      = Shader.PropertyToID("_BloomTint");
         public static readonly int _BloomTexture                   = Shader.PropertyToID("_BloomTexture");